abstracted out helper classes from test_evc, test_nextup -> contest_common.py
